This one took longer than expected, split user menu
