+++ demos/storybook/.storybook/main.js
@@ -11,11 +11,11 @@ module.exports = {
             name: '@storybook/addon-storysource',
             options: {
                 rule: {
-                    test: [/\.tsx?$/],
+                    // test: [/\.stories\.jsx?$/], This is default
                     include: [path.resolve(__dirname, '../stories')], // You can specify directories
                 },
                 loaderOptions: {
-                    prettierConfig: '@pxblue/prettier-config',
+                    prettierConfig: '@pxblue/prettier-config'
                 },
             },
         },
